nosql_optimized/postgres/ingestion/requirements.txt
requests
pandas
sqlalchemy
psycopg2-binary
python-dotenv

#pip install -r requirements.txt